Apply monkey patching only after loading a MySQL adapter.
