+++ config/initializers/ar_mysql_column_charset.rb
@@ -1,3 +1,1 @@
-ActiveSupport.on_load :active_record do
-  require 'ar_mysql_column_charset'
-end
+require 'ar_mysql_column_charset'
